jobboardsystem/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
cloudinary==1.44.1
Django==6.0.3
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-filter==25.2
django-js-asset==3.1.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.11
mysqlclient==2.2.8
pillow==12.1.1
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
